+++ src/renderer/src/components/diff-comments/decorated-diff-comment.ts
@@ -1,0 +1,10 @@
+import type { DiffComment } from '../../../../shared/diff-comment-types'
+
+export type DecoratedDiffComment = DiffComment & {
+  author?: string
+  authorAvatarUrl?: string
+  createdAtLabel?: string
+  url?: string
+  canDelete?: boolean
+  canEdit?: boolean
+}
