+++ src/renderer/src/components/diff-comments/useDiffCommentDecorator.tsx
@@ -1,30 +1,19 @@
-/* eslint-disable max-lines -- Why: this hook owns Monaco view-zone lifecycle, inline React roots, range selection, and scroll coordination in one place. */
 import { useCallback, useEffect, useMemo, useRef } from 'react'
-import * as monaco from 'monaco-editor'
 import type { editor as monacoEditor, IDisposable } from 'monaco-editor'
 import { createRoot, type Root } from 'react-dom/client'
-import type { DiffComment } from '../../../../shared/diff-comment-types'
 import { getCommentBodyLayoutLineCount } from '@/lib/comment-body-line-count'
-import { getDiffCommentLineLabel } from '@/lib/diff-comment-compat'
-import { formatDiffComments } from '@/lib/diff-comments-format'
 import { useAppStore } from '@/store'
-import { TooltipProvider } from '@/components/ui/tooltip'
-import { DiffCommentCard } from './DiffCommentCard'
-import { getDiffCommentPopoverTop } from './diff-comment-popover-position'
+import { installDiffCommentAddButtonOverlay } from './diff-comment-add-button-overlay'
 import { installDiffCommentZoneMouseDownStopper } from './diff-comment-zone-mouse-events'
-import { NotesSendMenu, type NotesSendMenuScope } from '../editor/NotesSendMenu'
-import { translate } from '@/i18n/i18n'
-
-// Monaco glyph decorations don't expose usable click events, so we own an absolutely-positioned "+" button that follows the hovered line.
-
-export type DecoratedDiffComment = DiffComment & {
-  author?: string
-  authorAvatarUrl?: string
-  createdAtLabel?: string
-  url?: string
-  canDelete?: boolean
-  canEdit?: boolean
-}
+import { getRenderSignature, renderDiffCommentZoneCard } from './diff-comment-zone-card'
+import type { DecoratedDiffComment } from './decorated-diff-comment'
+import {
+  resizeDiffCommentZone,
+  ZONE_CHROME_PX,
+  ZONE_LINE_PX,
+  ZONE_MIN_PX,
+  type ZoneEntry
+} from './diff-comment-view-zone-entry'
 
 type DecoratorArgs = {
   editor: monacoEditor.ICodeEditor | null
@@ -43,57 +32,6 @@ type DecoratorArgs = {
   // Pending scroll-to-note id from the sidebar; decorator reveals the line and acks so the same id can be re-requested later.
   pendingScrollCommentId?: string | null
   onPendingScrollConsumed?: () => void
-}
-
-type ZoneEntry = {
-  zoneId: string
-  domNode: HTMLElement
-  // Hold the delegate so layoutZone re-reads our updated heightInPx — mutating the delegate is the supported way to resize a zone in place.
-  delegate: monacoEditor.IViewZone
-  root: Root
-  disposeMouseDownStopper: () => void
-  lastRenderSignature: string
-  // First onDomNodeTop = deterministic "zone laid out" signal; gates scroll-to-note instead of polling getTopForLineNumber.
-  laidOut: boolean
-}
-
-// Card chrome + per-line body height; used by the initial estimate and the live resize, so keep in lockstep.
-const ZONE_CHROME_PX = 68
-const ZONE_LINE_PX = 20
-const ZONE_MIN_PX = 88
-
-function getRenderSignature(
-  comment: DecoratedDiffComment,
-  formatCommentPrompt?: (comment: DecoratedDiffComment) => string
-): string {
-  return JSON.stringify({
-    body: comment.body,
-    sentAt: comment.sentAt ?? null,
-    author: comment.author ?? null,
-    authorAvatarUrl: comment.authorAvatarUrl ?? null,
-    createdAtLabel: comment.createdAtLabel ?? null,
-    url: comment.url ?? null,
-    canDelete: comment.canDelete ?? null,
-    canEdit: comment.canEdit ?? null,
-    sendPrompt: formatCommentPrompt ? formatCommentPrompt(comment) : null
-  })
-}
-
-function getSingleCommentSendScopes(
-  comment: DecoratedDiffComment,
-  formatCommentPrompt?: (comment: DecoratedDiffComment) => string
-): NotesSendMenuScope<DecoratedDiffComment>[] {
-  return [
-    {
-      id: 'note',
-      label: translate(
-        'auto.components.diff.comments.useDiffCommentDecorator.995fa28b50',
-        'This note'
-      ),
-      notes: comment.sentAt ? [] : [comment],
-      prompt: formatCommentPrompt ? formatCommentPrompt(comment) : formatDiffComments([comment])
-    }
-  ]
 }
 
 export function useDiffCommentDecorator({
@@ -158,185 +96,18 @@ export function useDiffCommentDecorator({
     }
 
     const zones = zonesRef.current
-    const plus = document.createElement('button')
-    plus.type = 'button'
-    plus.className = 'orca-diff-comment-add-btn'
-    plus.title = addButtonLabel
-    plus.setAttribute('aria-label', addButtonLabel)
-    plus.innerHTML =
-      '<svg viewBox="0 0 16 16" width="12" height="12" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round"><path d="M8 3v10M3 8h10"/></svg>'
-    plus.style.display = 'none'
-    editorDomNode.appendChild(plus)
-
-    const getLineHeight = (): number => {
-      const h = editor.getOption(monaco.editor.EditorOption.lineHeight)
-      return typeof h === 'number' && h > 0 ? h : 19
-    }
-
-    // Cache last-applied styles so positionAtLine skips redundant DOM writes on high-freq mousemove (restyling under the cursor flickers).
-    let lastTop: number | null = null
-    let lastDisplay: string | null = null
-
-    const setDisplay = (value: string): void => {
-      if (lastDisplay === value) {
-        return
-      }
-      plus.style.display = value
-      lastDisplay = value
-    }
-
-    // Fixed 18px square centered in the line box — tracking line-height made a rectangle on taller line-heights.
-    const BUTTON_SIZE = 18
-    let rangeDecorationIds: string[] = []
-    let dragState: { startLine: number; endLine: number } | null = null
-
-    const clearRangeDecoration = (): void => {
-      if (rangeDecorationIds.length > 0) {
-        rangeDecorationIds = editor.deltaDecorations(rangeDecorationIds, [])
-      }
-    }
-
-    const updateRangeDecoration = (startLine: number, endLine: number): void => {
-      const from = Math.min(startLine, endLine)
-      const to = Math.max(startLine, endLine)
-      rangeDecorationIds = editor.deltaDecorations(rangeDecorationIds, [
-        {
-          range: new monaco.Range(from, 1, to, 1),
-          options: {
-            isWholeLine: true,
-            className: 'orca-diff-comment-range-highlight'
-          }
-        }
-      ])
-    }
-
-    const getLineAtClientPoint = (clientX: number, clientY: number): number | null => {
-      return editor.getTargetAtClientPoint(clientX, clientY)?.position?.lineNumber ?? null
-    }
-
-    const canCommentOnLine = (lineNumber: number): boolean => {
-      return commentableLineSet === null || commentableLineSet.has(lineNumber)
-    }
-
-    const canCommentOnRange = (startLine: number, endLine: number): boolean => {
-      if (commentableLineSet === null) {
-        return true
-      }
-      const from = Math.min(startLine, endLine)
-      const to = Math.max(startLine, endLine)
-      for (let line = from; line <= to; line++) {
-        if (!commentableLineSet.has(line)) {
-          return false
-        }
-      }
-      return true
-    }
-
-    const positionAtLine = (lineNumber: number): void => {
-      const lineTop = editor.getTopForLineNumber(lineNumber) - editor.getScrollTop()
-      const top = Math.round(lineTop + (getLineHeight() - BUTTON_SIZE) / 2)
-      if (top !== lastTop) {
-        plus.style.top = `${top}px`
-        lastTop = top
-      }
-      setDisplay('flex')
-    }
-
-    const finishRangeDrag = (ev: MouseEvent): void => {
-      ev.preventDefault()
-      ev.stopPropagation()
-      document.removeEventListener('mousemove', handleRangeDragMove)
-      document.removeEventListener('mouseup', finishRangeDrag)
-      const currentDrag = dragState
-      dragState = null
-      clearRangeDecoration()
-      if (!currentDrag) {
-        return
-      }
-      if (!canCommentOnRange(currentDrag.startLine, currentDrag.endLine)) {
-        return
-      }
-      const startLine = Math.min(currentDrag.startLine, currentDrag.endLine)
-      const lineNumber = Math.max(currentDrag.startLine, currentDrag.endLine)
-      const top = getDiffCommentPopoverTop(editor, lineNumber, getLineHeight())
-      if (top == null) {
-        return
-      }
-      onAddCommentClickRef.current({
-        lineNumber,
-        startLine: startLine === lineNumber ? undefined : startLine,
-        top
-      })
-    }
-
-    const handleRangeDragMove = (ev: MouseEvent): void => {
-      if (!dragState) {
-        return
-      }
-      const line = getLineAtClientPoint(ev.clientX, ev.clientY)
-      if (
-        line == null ||
-        line === dragState.endLine ||
-        !canCommentOnLine(line) ||
-        !canCommentOnRange(dragState.startLine, line)
-      ) {
-        return
-      }
-      dragState = { ...dragState, endLine: line }
-      updateRangeDecoration(dragState.startLine, line)
-    }
-
-    const handleMouseDown = (ev: MouseEvent): void => {
-      ev.preventDefault()
-      ev.stopPropagation()
-      const line = hoverLineRef.current
-      if (line == null || !canCommentOnLine(line)) {
-        return
-      }
-      dragState = { startLine: line, endLine: line }
-      updateRangeDecoration(line, line)
-      document.addEventListener('mousemove', handleRangeDragMove)
-      document.addEventListener('mouseup', finishRangeDrag)
-    }
-    plus.addEventListener('mousedown', handleMouseDown)
-
-    const onMouseMove = editor.onMouseMove((e) => {
-      // Monaco reports null position over our "+" button; hiding on null would flicker-loop, so keep it visible while the cursor's on it.
-      const srcEvent = e.event?.browserEvent as MouseEvent | undefined
-      if (srcEvent && plus.contains(srcEvent.target as Node)) {
-        return
-      }
-      const ln = e.target.position?.lineNumber ?? null
-      if (ln == null || !canCommentOnLine(ln)) {
-        hoverLineRef.current = null
-        setDisplay('none')
-        return
-      }
-      hoverLineRef.current = ln
-      positionAtLine(ln)
+    const disposeAddButtonOverlay = installDiffCommentAddButtonOverlay({
+      editor,
+      editorDomNode,
+      addButtonLabel,
+      commentableLineSet,
+      hoverLineRef,
+      disposablesRef,
+      onAddCommentClickRef
     })
-    // Keep hoverLineRef on mouse-leave: Monaco's content-area leave fires before the button's, so a click in that gap still resolves to the last-hovered line.
-    const onMouseLeave = editor.onMouseLeave(() => {
-      setDisplay('none')
-    })
-    const onScroll = editor.onDidScrollChange(() => {
-      if (hoverLineRef.current != null) {
-        positionAtLine(hoverLineRef.current)
-      }
-    })
-
-    disposablesRef.current = [onMouseMove, onMouseLeave, onScroll]
 
     return () => {
-      for (const d of disposablesRef.current) {
-        d.dispose()
-      }
-      disposablesRef.current = []
-      document.removeEventListener('mousemove', handleRangeDragMove)
-      document.removeEventListener('mouseup', finishRangeDrag)
-      clearRangeDecoration()
-      plus.removeEventListener('mousedown', handleMouseDown)
-      plus.remove()
+      disposeAddButtonOverlay()
       // Editor swapped/torn down: unmount roots and clear tracking so the next mount starts known-empty.
       // Defer unmount via queueMicrotask: a sync unmount during React's commit triggers React 19's "unmount while rendering" warning; clear zones synchronously.
       const rootsToUnmount = Array.from(zones.values(), (z) => {
@@ -370,31 +141,12 @@ export function useDiffCommentDecorator({
     // Unmounting a root inside changeViewZones races Monaco's zone bookkeeping; collect roots and unmount after the batch.
     const rootsToUnmount: Root[] = []
 
-    // Re-measure/re-layout the zone: mutate delegate.heightInPx first (Monaco's _layoutZone re-reads it) so inline edit expands without clipping.
     const resizeZone = (commentId: string): void => {
       const entry = zones.get(commentId)
       if (!entry) {
         return
       }
-      const child = entry.domNode.firstElementChild
-      const wrapperStyle = window.getComputedStyle(entry.domNode)
-      const verticalPadding =
-        Number.parseFloat(wrapperStyle.paddingTop) + Number.parseFloat(wrapperStyle.paddingBottom)
-      // Monaco pins the zone node to its previous height (scrollHeight can't shrink), so measure the rendered card+padding to allow collapse.
-      const childHeight = child?.getBoundingClientRect().height ?? 0
-      // React can commit while Monaco's zone is detached; preserve the safe
-      // initial estimate until the observer sees a measurable card.
-      if (childHeight <= 0) {
-        return
-      }
-      const measured = Math.ceil(childHeight + verticalPadding)
-      if (entry.delegate.heightInPx === measured) {
-        return
-      }
-      entry.delegate.heightInPx = measured
-      editor.changeViewZones((acc) => {
-        acc.layoutZone(entry.zoneId)
-      })
+      resizeDiffCommentZone(editor, entry)
     }
 
     // One-shot scroll resolver: getTopForLineNumber(line, includeZones=true) centers on the line+card pair (card sits in a zone above the line).
@@ -421,51 +173,16 @@ export function useDiffCommentDecorator({
 
     // Shared by the new-zone and patch branches so the card's prop wiring stays in lockstep.
     const renderCard = (root: Root, comment: DecoratedDiffComment): void => {
-      root.render(
-        // View zones are separate React roots outside the app root, so App.tsx context providers don't reach them.
-        <TooltipProvider delayDuration={400}>
-          <DiffCommentCard
-            lineNumber={comment.lineNumber}
-            startLine={comment.startLine}
-            label={comment.author ? getDiffCommentLineLabel(comment).toLowerCase() : undefined}
-            body={comment.body}
-            sentAt={comment.sentAt}
-            author={comment.author}
-            createdAtLabel={comment.createdAtLabel}
-            url={comment.url}
-            onDelete={
-              comment.canDelete === false ? undefined : () => onDeleteCommentRef.current(comment.id)
-            }
-            onSubmitEdit={
-              onUpdateCommentRef.current && comment.canEdit !== false
-                ? async (body) => {
-                    const fn = onUpdateCommentRef.current
-                    if (!fn) {
-                      return false
-                    }
-                    return fn(comment.id, body)
-                  }
-                : undefined
-            }
-            onContentResize={() => resizeZone(comment.id)}
-            observeRenderedSize
-            headerActions={
-              worktreeId && comment.author === undefined ? (
-                <NotesSendMenu
-                  worktreeId={worktreeId}
-                  groupId={activeGroupId}
-                  modeIdParts={['diff-comment-note', worktreeId, filePath, comment.id]}
-                  scopes={getSingleCommentSendScopes(comment, formatCommentPrompt)}
-                  targetModeLabel="This note"
-                  triggerClassName="orca-diff-comment-edit"
-                  disabledTooltip="Note already sent"
-                  onDelivered={(notes) => void clearDeliveredDiffComments(worktreeId, notes)}
-                />
-              ) : null
-            }
-          />
-        </TooltipProvider>
-      )
+      renderDiffCommentZoneCard(root, comment, {
+        worktreeId,
+        filePath,
+        activeGroupId,
+        formatCommentPrompt,
+        resizeZone,
+        onDeleteCommentRef,
+        onUpdateCommentRef,
+        clearDeliveredDiffComments
+      })
     }
 
     editor.changeViewZones((accessor) => {
