+++ src/renderer/src/components/editor/IpynbViewer.tsx
@@ -66,14 +66,11 @@ import {
   deleteIpynbCell,
   insertIpynbCell,
   moveIpynbCell,
-  parseIpynb,
   updateIpynbCellKind,
   updateIpynbCellOutputs,
-  updateIpynbCellSources,
-  type IpynbCell,
-  type IpynbCellKind,
-  type IpynbOutputItem
-} from './ipynb-parse'
+  updateIpynbCellSources
+} from './ipynb-cell-mutations'
+import { parseIpynb, type IpynbCell, type IpynbCellKind, type IpynbOutputItem } from './ipynb-parse'
 import { translate } from '@/i18n/i18n'
 
 type IpynbViewerProps = {
