+++ src/renderer/src/components/editor/editor-save-queue.ts
@@ -1,0 +1,237 @@
+import type { StoreApi } from 'zustand'
+import type { AppState } from '@/store'
+import type { OpenFile } from '@/store/slices/editor'
+import { findWorktreeById } from '@/store/slices/worktree-helpers'
+import { writeRuntimeFile } from '@/runtime/runtime-file-client'
+import { getEditorFileOperationContext } from '@/lib/editor-file-operation-owner'
+import {
+  canAutoSaveOpenFile,
+  isAutosaveSuspendedForFile,
+  normalizeAutoSaveDelayMs,
+  ORCA_EDITOR_FILE_SAVED_EVENT,
+  type EditorFileSavedDetail
+} from './editor-autosave'
+import { flushPendingEditorChange } from './editor-pending-flush'
+import {
+  clearSelfWrite,
+  recordSelfWrite,
+  SELF_WRITE_REMOTE_TTL_MS
+} from './editor-self-write-registry'
+import { getDiskBaselineSignature } from './diff-content-signature'
+import { trackExternalChangeConflictAction } from './editor-external-change-telemetry'
+
+export type AppStoreApi = Pick<StoreApi<AppState>, 'getState' | 'subscribe'>
+
+export type EditorSaveQueue = {
+  queueSave: (
+    file: OpenFile,
+    fallbackContent: string,
+    trigger?: 'autosave' | 'user'
+  ) => Promise<void>
+  quiesceFileSave: (fileId: string) => Promise<void>
+  clearAutoSaveTimer: (fileId: string) => void
+  bumpSaveGeneration: (fileId: string) => void
+  syncAutoSave: () => void
+  dispose: () => void
+}
+
+// Why: keeping the save queue, quiesce coordination, and the debounce timers that feed it together avoids split-brain saves.
+export function createEditorSaveQueue(store: AppStoreApi): EditorSaveQueue {
+  const autoSaveTimers = new Map<string, number>()
+  const autoSaveScheduledContent = new Map<string, string>()
+  const saveQueue = new Map<string, Promise<void>>()
+  const saveGeneration = new Map<string, number>()
+
+  const clearAutoSaveTimer = (fileId: string): void => {
+    const timerId = autoSaveTimers.get(fileId)
+    if (timerId !== undefined) {
+      window.clearTimeout(timerId)
+      autoSaveTimers.delete(fileId)
+    }
+    autoSaveScheduledContent.delete(fileId)
+  }
+
+  const bumpSaveGeneration = (fileId: string): void => {
+    saveGeneration.set(fileId, (saveGeneration.get(fileId) ?? 0) + 1)
+  }
+
+  const queueSave = (
+    file: OpenFile,
+    fallbackContent: string,
+    trigger: 'autosave' | 'user' = 'user'
+  ): Promise<void> => {
+    clearAutoSaveTimer(file.id)
+    const queuedGeneration = saveGeneration.get(file.id) ?? 0
+
+    const previousSave = saveQueue.get(file.id) ?? Promise.resolve()
+    const queuedSave = previousSave
+      .catch(() => undefined)
+      .then(async () => {
+        if ((saveGeneration.get(file.id) ?? 0) !== queuedGeneration) {
+          return
+        }
+
+        const state = store.getState()
+        const liveFile = state.openFiles.find((openFile) => openFile.id === file.id) ?? null
+        if (!liveFile) {
+          return
+        }
+
+        // Why: read-only tabs (AI Vault View Log) must never write the agent-owned artifact through editor paths.
+        if (liveFile.readOnly === true) {
+          return
+        }
+
+        if (liveFile.pendingOwnerMigration === true) {
+          if (trigger === 'autosave') {
+            return
+          }
+          throw new Error('This file is still restoring its workspace owner. Try saving again.')
+        }
+
+        // Why: only autosave is blocked while suspended; explicit user saves proceed (the banner warned).
+        if (trigger === 'autosave' && isAutosaveSuspendedForFile(liveFile)) {
+          return
+        }
+
+        const contentToSave = state.editorDrafts[file.id] ?? fallbackContent
+        const worktree = liveFile.worktreeId
+          ? findWorktreeById(state.worktreesByRepo ?? {}, liveFile.worktreeId)
+          : null
+        const fileContext = getEditorFileOperationContext(state, liveFile, worktree?.path ?? null)
+        const connectionId = fileContext.connectionId
+        // Why: stamp before writing so useEditorExternalWatch ignores our own fs:changed echo (editor-self-write-registry).
+        recordSelfWrite(
+          liveFile.filePath,
+          contentToSave,
+          liveFile.runtimeEnvironmentId,
+          connectionId || liveFile.runtimeEnvironmentId?.trim()
+            ? SELF_WRITE_REMOTE_TTL_MS
+            : undefined
+        )
+        try {
+          await writeRuntimeFile(fileContext, liveFile.filePath, contentToSave)
+        } catch (error) {
+          // Why: the self-write stamp is only valid after a real write; clear on failure so it can't suppress a real update.
+          clearSelfWrite(liveFile.filePath, liveFile.runtimeEnvironmentId)
+          throw error
+        }
+
+        if ((saveGeneration.get(file.id) ?? 0) !== queuedGeneration) {
+          return
+        }
+
+        const nextState = store.getState()
+        const currentDraft = nextState.editorDrafts[file.id]
+        const stillDirty = currentDraft !== undefined && currentDraft !== contentToSave
+        nextState.markFileDirty(file.id, stillDirty)
+        if (!stillDirty) {
+          nextState.clearEditorDraft(file.id)
+        }
+        // Why: disk now holds contentToSave — rebaseline so our own save isn't flagged external; drop pending verification.
+        nextState.setLastKnownDiskSignature(file.id, getDiskBaselineSignature(contentToSave))
+        nextState.clearPendingDiskBaselineVerification(file.id)
+        // Why: the write made disk match the buffer, so clear any now-stale changed-on-disk conflict.
+        const savedFile = nextState.openFiles.find((openFile) => openFile.id === file.id)
+        if (savedFile?.externalMutation === 'changed') {
+          trackExternalChangeConflictAction(savedFile, 'save_overwrite')
+          nextState.setExternalMutation(file.id, null)
+        }
+
+        window.dispatchEvent(
+          new CustomEvent<EditorFileSavedDetail>(ORCA_EDITOR_FILE_SAVED_EVENT, {
+            detail: { fileId: file.id, content: contentToSave }
+          })
+        )
+      })
+
+    let trackedSave: Promise<void>
+    trackedSave = queuedSave.finally(() => {
+      if (saveQueue.get(file.id) === trackedSave) {
+        saveQueue.delete(file.id)
+      }
+    })
+    saveQueue.set(file.id, trackedSave)
+    return trackedSave
+  }
+
+  const quiesceFileSave = async (fileId: string): Promise<void> => {
+    // Why: rich markdown debounces serialization, so force the pending draft out before we cancel timers.
+    flushPendingEditorChange(fileId)
+    const pendingSave = saveQueue.get(fileId)
+    clearAutoSaveTimer(fileId)
+    bumpSaveGeneration(fileId)
+    await pendingSave?.catch(() => undefined)
+  }
+
+  const syncAutoSave = (): void => {
+    const state = store.getState()
+    const openFilesById = new Map(state.openFiles.map((file) => [file.id, file]))
+
+    for (const fileId of Array.from(autoSaveTimers.keys())) {
+      const file = openFilesById.get(fileId)
+      const draft = state.editorDrafts[fileId]
+      const shouldKeepTimer =
+        state.settings?.editorAutoSave &&
+        file &&
+        file.isDirty &&
+        canAutoSaveOpenFile(file) &&
+        // Why: suspension holds until the user picks a side via the banner (or saves manually).
+        !isAutosaveSuspendedForFile(file) &&
+        draft !== undefined
+      if (!shouldKeepTimer) {
+        clearAutoSaveTimer(fileId)
+      }
+    }
+
+    if (!state.settings?.editorAutoSave) {
+      return
+    }
+
+    const autoSaveDelayMs = normalizeAutoSaveDelayMs(state.settings.editorAutoSaveDelayMs)
+    for (const file of state.openFiles) {
+      const draft = state.editorDrafts[file.id]
+      if (
+        !file.isDirty ||
+        draft === undefined ||
+        !canAutoSaveOpenFile(file) ||
+        isAutosaveSuspendedForFile(file)
+      ) {
+        clearAutoSaveTimer(file.id)
+        continue
+      }
+
+      if (autoSaveTimers.has(file.id) && autoSaveScheduledContent.get(file.id) === draft) {
+        continue
+      }
+
+      clearAutoSaveTimer(file.id)
+      autoSaveScheduledContent.set(file.id, draft)
+      const timerId = window.setTimeout(() => {
+        autoSaveTimers.delete(file.id)
+        autoSaveScheduledContent.delete(file.id)
+        void queueSave(file, draft, 'autosave')
+      }, autoSaveDelayMs)
+      autoSaveTimers.set(file.id, timerId)
+    }
+  }
+
+  const dispose = (): void => {
+    for (const timerId of autoSaveTimers.values()) {
+      window.clearTimeout(timerId)
+    }
+    autoSaveTimers.clear()
+    autoSaveScheduledContent.clear()
+    saveQueue.clear()
+    saveGeneration.clear()
+  }
+
+  return {
+    queueSave,
+    quiesceFileSave,
+    clearAutoSaveTimer,
+    bumpSaveGeneration,
+    syncAutoSave,
+    dispose
+  }
+}
