+++ src/renderer/src/lib/file-type-icon-extension-table.ts
@@ -1,0 +1,200 @@
+import {
+  Database,
+  FileArchive,
+  FileAxis3D,
+  FileBraces,
+  FileChartColumn,
+  FileCode,
+  FileCog,
+  FileDiff,
+  FileImage,
+  FileJson,
+  FileKey,
+  FileLock,
+  FileMusic,
+  FileSliders,
+  FileSpreadsheet,
+  FileTerminal,
+  FileText,
+  FileType,
+  FileVideo,
+  type LucideIcon
+} from 'lucide-react'
+
+export const FILE_ICON_BY_EXTENSION: Record<string, LucideIcon> = {
+  '7z': FileArchive,
+  aac: FileMusic,
+  adoc: FileText,
+  ai: FileImage,
+  asc: FileKey,
+  astro: FileCode,
+  avi: FileVideo,
+  avif: FileImage,
+  bash: FileTerminal,
+  bat: FileTerminal,
+  blend: FileAxis3D,
+  bmp: FileImage,
+  br: FileArchive,
+  bz2: FileArchive,
+  c: FileCode,
+  cc: FileCode,
+  cer: FileKey,
+  cfg: FileSliders,
+  cjs: FileCode,
+  clj: FileCode,
+  cmd: FileTerminal,
+  conf: FileSliders,
+  cpp: FileCode,
+  crt: FileKey,
+  cs: FileCode,
+  css: FileType,
+  csv: FileSpreadsheet,
+  cts: FileCode,
+  cxx: FileCode,
+  dart: FileCode,
+  db: Database,
+  diff: FileDiff,
+  dmg: FileArchive,
+  doc: FileText,
+  docx: FileText,
+  duckdb: Database,
+  eot: FileType,
+  eps: FileImage,
+  erl: FileCode,
+  ex: FileCode,
+  exs: FileCode,
+  fbx: FileAxis3D,
+  fish: FileTerminal,
+  flac: FileMusic,
+  fs: FileCode,
+  fsx: FileCode,
+  gif: FileImage,
+  glb: FileAxis3D,
+  gltf: FileAxis3D,
+  go: FileCode,
+  gpg: FileKey,
+  gql: FileBraces,
+  gradle: FileCog,
+  graphql: FileBraces,
+  gz: FileArchive,
+  h: FileCode,
+  hcl: FileSliders,
+  heic: FileImage,
+  hpp: FileCode,
+  hrl: FileCode,
+  hs: FileCode,
+  htm: FileCode,
+  html: FileCode,
+  ico: FileImage,
+  ini: FileSliders,
+  ipynb: FileChartColumn,
+  iso: FileArchive,
+  java: FileCode,
+  jpeg: FileImage,
+  jpg: FileImage,
+  js: FileCode,
+  json: FileJson,
+  json5: FileJson,
+  jsonc: FileJson,
+  jsx: FileCode,
+  key: FileKey,
+  kt: FileCode,
+  kts: FileCode,
+  less: FileType,
+  lock: FileLock,
+  log: FileText,
+  lua: FileCode,
+  m4a: FileMusic,
+  m4v: FileVideo,
+  md: FileText,
+  mdx: FileText,
+  mjs: FileCode,
+  mkv: FileVideo,
+  mmd: FileChartColumn,
+  mov: FileVideo,
+  mp3: FileMusic,
+  mp4: FileVideo,
+  mpeg: FileVideo,
+  mpg: FileVideo,
+  mts: FileCode,
+  nim: FileCode,
+  nu: FileTerminal,
+  obj: FileAxis3D,
+  ods: FileSpreadsheet,
+  ogg: FileMusic,
+  opus: FileMusic,
+  otf: FileType,
+  p12: FileLock,
+  patch: FileDiff,
+  pdf: FileText,
+  pem: FileKey,
+  pfx: FileLock,
+  php: FileCode,
+  pl: FileCode,
+  pm: FileCode,
+  png: FileImage,
+  ppt: FileChartColumn,
+  pptx: FileChartColumn,
+  prisma: Database,
+  properties: FileSliders,
+  proto: FileBraces,
+  ps1: FileTerminal,
+  psd: FileImage,
+  pub: FileKey,
+  py: FileCode,
+  r: FileCode,
+  rar: FileArchive,
+  rb: FileCode,
+  rst: FileText,
+  rs: FileCode,
+  rtf: FileText,
+  sass: FileType,
+  scala: FileCode,
+  scss: FileType,
+  sh: FileTerminal,
+  sol: FileCode,
+  sqlite: Database,
+  sqlite3: Database,
+  sql: Database,
+  stl: FileAxis3D,
+  svelte: FileCode,
+  svg: FileImage,
+  swift: FileCode,
+  tar: FileArchive,
+  'tar.bz2': FileArchive,
+  'tar.gz': FileArchive,
+  'tar.xz': FileArchive,
+  tbz2: FileArchive,
+  tex: FileText,
+  tf: FileSliders,
+  tfvars: FileSliders,
+  tgz: FileArchive,
+  tif: FileImage,
+  tiff: FileImage,
+  toml: FileSliders,
+  ts: FileCode,
+  tsx: FileCode,
+  tsv: FileSpreadsheet,
+  ttf: FileType,
+  txt: FileText,
+  txz: FileArchive,
+  vb: FileCode,
+  vue: FileCode,
+  wav: FileMusic,
+  webm: FileVideo,
+  webp: FileImage,
+  woff: FileType,
+  woff2: FileType,
+  xhtml: FileCode,
+  xls: FileSpreadsheet,
+  xlsx: FileSpreadsheet,
+  xml: FileCode,
+  xz: FileArchive,
+  yaml: FileSliders,
+  yml: FileSliders,
+  zig: FileCode,
+  zip: FileArchive,
+  zsh: FileTerminal
+}
+
+export const COMPOUND_EXTENSIONS = ['tar.bz2', 'tar.gz', 'tar.xz']
