+++ src/renderer/src/lib/file-type-icon-name-table.ts
@@ -1,0 +1,87 @@
+import {
+  FileBox,
+  FileCog,
+  FileJson,
+  FileKey,
+  FileLock,
+  FileSliders,
+  FileTerminal,
+  FileText,
+  type LucideIcon
+} from 'lucide-react'
+
+export const FILE_ICON_BY_NAME: Record<string, LucideIcon> = {
+  '.babelrc': FileSliders,
+  '.dockerignore': FileSliders,
+  '.editorconfig': FileSliders,
+  '.eslintrc': FileSliders,
+  '.eslintrc.cjs': FileSliders,
+  '.eslintrc.js': FileSliders,
+  '.eslintrc.json': FileJson,
+  '.eslintrc.yaml': FileSliders,
+  '.eslintrc.yml': FileSliders,
+  '.gitattributes': FileSliders,
+  '.gitignore': FileSliders,
+  '.npmrc': FileSliders,
+  '.prettierrc': FileSliders,
+  '.prettierrc.json': FileJson,
+  '.prettierrc.yaml': FileSliders,
+  '.prettierrc.yml': FileSliders,
+  'agents.md': FileText,
+  authors: FileText,
+  'bun.lock': FileBox,
+  'bun.lockb': FileBox,
+  'cargo.lock': FileBox,
+  'cargo.toml': FileBox,
+  changelog: FileText,
+  'changelog.md': FileText,
+  'cmakelists.txt': FileCog,
+  codeowners: FileKey,
+  'components.json': FileSliders,
+  'composer.json': FileBox,
+  'composer.lock': FileBox,
+  contributing: FileText,
+  'contributing.md': FileText,
+  copying: FileKey,
+  dockerfile: FileCog,
+  gemfile: FileBox,
+  'go.mod': FileBox,
+  'go.sum': FileBox,
+  license: FileKey,
+  makefile: FileTerminal,
+  'meson.build': FileCog,
+  notice: FileKey,
+  'package-lock.json': FileBox,
+  'package.json': FileBox,
+  pipfile: FileBox,
+  'pnpm-lock.yaml': FileBox,
+  'pnpm-workspace.yaml': FileBox,
+  'poetry.lock': FileBox,
+  'pom.xml': FileBox,
+  'postcss.config.cjs': FileSliders,
+  'postcss.config.js': FileSliders,
+  'postcss.config.mjs': FileSliders,
+  'postcss.config.ts': FileSliders,
+  'pyproject.toml': FileBox,
+  readme: FileText,
+  'readme.md': FileText,
+  'requirements-dev.txt': FileBox,
+  'requirements.txt': FileBox,
+  security: FileLock,
+  'security.md': FileLock,
+  'settings.gradle': FileCog,
+  'settings.gradle.kts': FileCog,
+  'tailwind.config.cjs': FileSliders,
+  'tailwind.config.js': FileSliders,
+  'tailwind.config.mjs': FileSliders,
+  'tailwind.config.ts': FileSliders,
+  todo: FileText,
+  'tsconfig.json': FileSliders,
+  'vite.config.js': FileSliders,
+  'vite.config.mjs': FileSliders,
+  'vite.config.ts': FileSliders,
+  'vitest.config.js': FileSliders,
+  'vitest.config.mjs': FileSliders,
+  'vitest.config.ts': FileSliders,
+  'yarn.lock': FileBox
+}
