+++ src/renderer/src/lib/file-type-icons.ts
@@ -1,283 +1,6 @@
-/* eslint-disable max-lines -- Why: this module is intentionally a compact filename/extension icon table. */
-import {
-  Database,
-  File,
-  FileArchive,
-  FileAxis3D,
-  FileBox,
-  FileBraces,
-  FileChartColumn,
-  FileCode,
-  FileCog,
-  FileDiff,
-  FileImage,
-  FileJson,
-  FileKey,
-  FileLock,
-  FileMusic,
-  FileSliders,
-  FileSpreadsheet,
-  FileTerminal,
-  FileText,
-  FileType,
-  FileVideo,
-  Smartphone,
-  type LucideIcon
-} from 'lucide-react'
-
-const FILE_ICON_BY_NAME: Record<string, LucideIcon> = {
-  '.babelrc': FileSliders,
-  '.dockerignore': FileSliders,
-  '.editorconfig': FileSliders,
-  '.eslintrc': FileSliders,
-  '.eslintrc.cjs': FileSliders,
-  '.eslintrc.js': FileSliders,
-  '.eslintrc.json': FileJson,
-  '.eslintrc.yaml': FileSliders,
-  '.eslintrc.yml': FileSliders,
-  '.gitattributes': FileSliders,
-  '.gitignore': FileSliders,
-  '.npmrc': FileSliders,
-  '.prettierrc': FileSliders,
-  '.prettierrc.json': FileJson,
-  '.prettierrc.yaml': FileSliders,
-  '.prettierrc.yml': FileSliders,
-  'agents.md': FileText,
-  authors: FileText,
-  'bun.lock': FileBox,
-  'bun.lockb': FileBox,
-  'cargo.lock': FileBox,
-  'cargo.toml': FileBox,
-  changelog: FileText,
-  'changelog.md': FileText,
-  'cmakelists.txt': FileCog,
-  codeowners: FileKey,
-  'components.json': FileSliders,
-  'composer.json': FileBox,
-  'composer.lock': FileBox,
-  contributing: FileText,
-  'contributing.md': FileText,
-  copying: FileKey,
-  dockerfile: FileCog,
-  gemfile: FileBox,
-  'go.mod': FileBox,
-  'go.sum': FileBox,
-  license: FileKey,
-  makefile: FileTerminal,
-  'meson.build': FileCog,
-  notice: FileKey,
-  'package-lock.json': FileBox,
-  'package.json': FileBox,
-  pipfile: FileBox,
-  'pnpm-lock.yaml': FileBox,
-  'pnpm-workspace.yaml': FileBox,
-  'poetry.lock': FileBox,
-  'pom.xml': FileBox,
-  'postcss.config.cjs': FileSliders,
-  'postcss.config.js': FileSliders,
-  'postcss.config.mjs': FileSliders,
-  'postcss.config.ts': FileSliders,
-  'pyproject.toml': FileBox,
-  readme: FileText,
-  'readme.md': FileText,
-  'requirements-dev.txt': FileBox,
-  'requirements.txt': FileBox,
-  security: FileLock,
-  'security.md': FileLock,
-  'settings.gradle': FileCog,
-  'settings.gradle.kts': FileCog,
-  'tailwind.config.cjs': FileSliders,
-  'tailwind.config.js': FileSliders,
-  'tailwind.config.mjs': FileSliders,
-  'tailwind.config.ts': FileSliders,
-  todo: FileText,
-  'tsconfig.json': FileSliders,
-  'vite.config.js': FileSliders,
-  'vite.config.mjs': FileSliders,
-  'vite.config.ts': FileSliders,
-  'vitest.config.js': FileSliders,
-  'vitest.config.mjs': FileSliders,
-  'vitest.config.ts': FileSliders,
-  'yarn.lock': FileBox
-}
-
-const FILE_ICON_BY_EXTENSION: Record<string, LucideIcon> = {
-  '7z': FileArchive,
-  aac: FileMusic,
-  adoc: FileText,
-  ai: FileImage,
-  asc: FileKey,
-  astro: FileCode,
-  avi: FileVideo,
-  avif: FileImage,
-  bash: FileTerminal,
-  bat: FileTerminal,
-  blend: FileAxis3D,
-  bmp: FileImage,
-  br: FileArchive,
-  bz2: FileArchive,
-  c: FileCode,
-  cc: FileCode,
-  cer: FileKey,
-  cfg: FileSliders,
-  cjs: FileCode,
-  clj: FileCode,
-  cmd: FileTerminal,
-  conf: FileSliders,
-  cpp: FileCode,
-  crt: FileKey,
-  cs: FileCode,
-  css: FileType,
-  csv: FileSpreadsheet,
-  cts: FileCode,
-  cxx: FileCode,
-  dart: FileCode,
-  db: Database,
-  diff: FileDiff,
-  dmg: FileArchive,
-  doc: FileText,
-  docx: FileText,
-  duckdb: Database,
-  eot: FileType,
-  eps: FileImage,
-  erl: FileCode,
-  ex: FileCode,
-  exs: FileCode,
-  fbx: FileAxis3D,
-  fish: FileTerminal,
-  flac: FileMusic,
-  fs: FileCode,
-  fsx: FileCode,
-  gif: FileImage,
-  glb: FileAxis3D,
-  gltf: FileAxis3D,
-  go: FileCode,
-  gpg: FileKey,
-  gql: FileBraces,
-  gradle: FileCog,
-  graphql: FileBraces,
-  gz: FileArchive,
-  h: FileCode,
-  hcl: FileSliders,
-  heic: FileImage,
-  hpp: FileCode,
-  hrl: FileCode,
-  hs: FileCode,
-  htm: FileCode,
-  html: FileCode,
-  ico: FileImage,
-  ini: FileSliders,
-  ipynb: FileChartColumn,
-  iso: FileArchive,
-  java: FileCode,
-  jpeg: FileImage,
-  jpg: FileImage,
-  js: FileCode,
-  json: FileJson,
-  json5: FileJson,
-  jsonc: FileJson,
-  jsx: FileCode,
-  key: FileKey,
-  kt: FileCode,
-  kts: FileCode,
-  less: FileType,
-  lock: FileLock,
-  log: FileText,
-  lua: FileCode,
-  m4a: FileMusic,
-  m4v: FileVideo,
-  md: FileText,
-  mdx: FileText,
-  mjs: FileCode,
-  mkv: FileVideo,
-  mmd: FileChartColumn,
-  mov: FileVideo,
-  mp3: FileMusic,
-  mp4: FileVideo,
-  mpeg: FileVideo,
-  mpg: FileVideo,
-  mts: FileCode,
-  nim: FileCode,
-  nu: FileTerminal,
-  obj: FileAxis3D,
-  ods: FileSpreadsheet,
-  ogg: FileMusic,
-  opus: FileMusic,
-  otf: FileType,
-  p12: FileLock,
-  patch: FileDiff,
-  pdf: FileText,
-  pem: FileKey,
-  pfx: FileLock,
-  php: FileCode,
-  pl: FileCode,
-  pm: FileCode,
-  png: FileImage,
-  ppt: FileChartColumn,
-  pptx: FileChartColumn,
-  prisma: Database,
-  properties: FileSliders,
-  proto: FileBraces,
-  ps1: FileTerminal,
-  psd: FileImage,
-  pub: FileKey,
-  py: FileCode,
-  r: FileCode,
-  rar: FileArchive,
-  rb: FileCode,
-  rst: FileText,
-  rs: FileCode,
-  rtf: FileText,
-  sass: FileType,
-  scala: FileCode,
-  scss: FileType,
-  sh: FileTerminal,
-  sol: FileCode,
-  sqlite: Database,
-  sqlite3: Database,
-  sql: Database,
-  stl: FileAxis3D,
-  svelte: FileCode,
-  svg: FileImage,
-  swift: FileCode,
-  tar: FileArchive,
-  'tar.bz2': FileArchive,
-  'tar.gz': FileArchive,
-  'tar.xz': FileArchive,
-  tbz2: FileArchive,
-  tex: FileText,
-  tf: FileSliders,
-  tfvars: FileSliders,
-  tgz: FileArchive,
-  tif: FileImage,
-  tiff: FileImage,
-  toml: FileSliders,
-  ts: FileCode,
-  tsx: FileCode,
-  tsv: FileSpreadsheet,
-  ttf: FileType,
-  txt: FileText,
-  txz: FileArchive,
-  vb: FileCode,
-  vue: FileCode,
-  wav: FileMusic,
-  webm: FileVideo,
-  webp: FileImage,
-  woff: FileType,
-  woff2: FileType,
-  xhtml: FileCode,
-  xls: FileSpreadsheet,
-  xlsx: FileSpreadsheet,
-  xml: FileCode,
-  xz: FileArchive,
-  yaml: FileSliders,
-  yml: FileSliders,
-  zig: FileCode,
-  zip: FileArchive,
-  zsh: FileTerminal
-}
-
-const COMPOUND_EXTENSIONS = ['tar.bz2', 'tar.gz', 'tar.xz']
+import { File, FileCog, FileLock, FileTerminal, Smartphone, type LucideIcon } from 'lucide-react'
+import { COMPOUND_EXTENSIONS, FILE_ICON_BY_EXTENSION } from './file-type-icon-extension-table'
+import { FILE_ICON_BY_NAME } from './file-type-icon-name-table'
 
 function getFilename(filePath: string | undefined | null): string {
   if (!filePath) {
